fix(build): split json-render catalog to avoid React in server bundle
Move @json-render/react dependency into json-render-catalog-with-schema.ts
so the agent route only imports the plain string array (no React dependency).
This fixes "createContext is not a function" during Next.js static analysis.
