+++ components/agents/chat/message.tsx
@@ -42,8 +42,8 @@ import '../markdown-code.css'
 
 import { MermaidRenderer } from './mermaid-renderer'
 import { type AgentToolPart, ToolCallPart } from './tool-output'
+import { AGENT_JSON_RENDER_CATALOG } from '@/lib/ai/agent/json-render-catalog-with-schema'
 import {
-  AGENT_JSON_RENDER_CATALOG,
   AGENT_JSON_RENDER_MAX_ELEMENT_COUNT,
   AGENT_JSON_RENDER_MAX_SPEC_BYTES,
   AGENT_JSON_RENDER_MAX_SPEC_PART_BYTES,
