+++ lib/ai/agent/json-render-catalog.ts
@@ -1,7 +1,3 @@
-import { defineCatalog } from '@json-render/core'
-import { schema as reactSchema } from '@json-render/react'
-import { shadcnComponentDefinitions } from '@json-render/shadcn/catalog'
-
 export {
   AGENT_JSON_RENDER_MAX_ELEMENT_COUNT,
   AGENT_JSON_RENDER_MAX_SPEC_BYTES,
@@ -22,22 +18,3 @@ export const AGENT_JSON_RENDER_COMPONENT_NAMES = [
   'Skeleton',
   'Spinner',
 ] as const
-
-const AGENT_JSON_RENDER_COMPONENTS = {
-  Card: shadcnComponentDefinitions.Card,
-  Stack: shadcnComponentDefinitions.Stack,
-  Grid: shadcnComponentDefinitions.Grid,
-  Separator: shadcnComponentDefinitions.Separator,
-  Heading: shadcnComponentDefinitions.Heading,
-  Text: shadcnComponentDefinitions.Text,
-  Alert: shadcnComponentDefinitions.Alert,
-  Badge: shadcnComponentDefinitions.Badge,
-  Progress: shadcnComponentDefinitions.Progress,
-  Skeleton: shadcnComponentDefinitions.Skeleton,
-  Spinner: shadcnComponentDefinitions.Spinner,
-} as const
-
-export const AGENT_JSON_RENDER_CATALOG = defineCatalog(reactSchema, {
-  components: AGENT_JSON_RENDER_COMPONENTS,
-  actions: {},
-})
